moved BatchProfilesPyramid into separate file
