+++ packages/nextjs/app/page.tsx
@@ -1,12 +1,10 @@
 "use client";
 
-import Image from "next/image";
 import Link from "next/link";
-import { useTicTacToeState } from "./TicTacToeContext";
+import BatchProfilesPyramid from "./BatchProfilesPyramid";
 import TicTacToeGame from "./TicTacToeGame";
 import { BugAntIcon, MagnifyingGlassIcon } from "@heroicons/react/24/outline";
 import { useScaffoldReadContract } from "~~/hooks/scaffold-eth";
-import { images } from "~~/utils/scaffold-eth/batchMembersImages";
 
 function DetailsComponent(): JSX.Element {
   return (
@@ -15,57 +13,6 @@ function DetailsComponent(): JSX.Element {
 
       <TicTacToeGame />
       <BatchProfilesPyramid />
-    </div>
-  );
-}
-
-function BatchProfilesPyramid(): JSX.Element {
-  // Build pyramid levels
-  const pyramidLevels: string[][] = [];
-  const remainingImages: string[] = [...images];
-  let level: number = 1;
-  while (remainingImages.length > 0) {
-    const levelImages: string[] = remainingImages.splice(0, level);
-    pyramidLevels.push(levelImages);
-    level++;
-  }
-  const { activeImages, ticTacToeWon, showTicTacToe, setShowTicTacToe, setActiveImages } = useTicTacToeState();
-
-  function toggleImage(index: number): void {
-    if (ticTacToeWon) {
-      const newActiveImages: boolean[] = [...activeImages];
-      newActiveImages[index] = !newActiveImages[index];
-      setActiveImages(newActiveImages);
-    } else if (!showTicTacToe) {
-      setShowTicTacToe(true);
-    }
-  }
-
-  return (
-    <div className="relative w-full max-w-4xl mx-auto">
-      {pyramidLevels.map((levelImages: string[], levelIndex: number) => (
-        <div key={levelIndex} className="flex justify-center items-center mb-16 relative">
-          {levelImages.map((img: string, i: number) => {
-            const globalIndex: number = images.indexOf(img);
-            const isActive: boolean = activeImages[globalIndex];
-            return (
-              <div key={img + i} className="relative mx-4">
-                <Image
-                  src={img}
-                  alt={`Image ${globalIndex + 1}`}
-                  width={485}
-                  height={485}
-                  className={`w-32 h-32 object-cover rounded-lg cursor-pointer transition-all duration-300 ${
-                    isActive ? "opacity-100" : "opacity-30 grayscale"
-                  }`}
-                  onClick={() => toggleImage(globalIndex)}
-                  priority={globalIndex < 3}
-                />
-              </div>
-            );
-          })}
-        </div>
-      ))}
     </div>
   );
 }
